+++ src/components/NewEntity/NewEntity.tsx
@@ -17,7 +17,7 @@ import checkName from '@helpers/checkName'
 import ShortcutWidget from '@components/ShortcutWidget'
 import { useSelectionContext, useProjectTableContext } from '@shared/ProjectTreeTable'
 import { parseCellId } from '@shared/ProjectTreeTable/utils/cellUtils'
-import { EditorTaskNode, MatchingFolder } from '@shared/ProjectTreeTable/utils/types'
+import { EditorTaskNode, MatchingFolder } from '@shared/ProjectTreeTable'
 import FolderSequence from '@components/FolderSequence/FolderSequence'
 import { EntityForm, NewEntityType, useNewEntityContext } from '@context/NewEntityContext'
 import { ProjectModel } from '@api/rest/project'
