+++ src/shared/src/ContextMenu/index.ts
@@ -1,4 +1,4 @@
-import { useCreateContextMenu } from './useCreateContextMenu'
-export { useCreateContextMenu } from './useCreateContextMenu'
-export type { ContextMenuItemType } from './useCreateContextMenu'
-export default useCreateContextMenu
+export * from './ContextMenuContext'
+export * from './ContextMenuItem'
+export * from './GlobalContextMenu'
+export * from './useCreateContextMenu'
