+++ src/shared/src/ProjectTreeTable/utils/index.ts
@@ -1,2 +1,3 @@
 export * from '../types/table'
 export * from './clientFilterToQueryFilter'
+export * from './loadingUtils'
