+++ src/shared/src/index.tsx
@@ -1,2 +1,5 @@
 export * from './ProjectTreeTable'
 export * from './hooks'
+export * from './helpers'
+export * from './EmptyPlaceholder'
+export * from './ContextMenu'
